backend/direct_chat_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
langchain>=0.1.0
langchain-community>=0.0.10
requests==2.31.0
python-multipart==0.0.6
typing-extensions==4.8.0
httpx==0.25.0
PyYAML==6.0.1
